requirements.txt
discord
asyncio
pillow
requests
pymongo
typing
